refactor wms function to libeli, add more debug information
