+++ requirements-wms_sync.txt
@@ -1,11 +1,11 @@
-pillow==8.1.0
-aiohttp==3.7.4
+pillow==8.1.2
+aiohttp==3.7.4.post0
 aiofiles==0.6.0
 imagehash==4.2.0
 mercantile==1.1.6
 shapely==1.7.1
-pyproj==3.0.0.post1
+pyproj==3.0.1
 jsonschema==3.2.0
 colorlog==4.7.2
-pyproj==3.0.0.post1
-python-magic==0.4.18
+pyproj==3.0.1
+python-magic==0.4.22
